relocate Suicide Detector and StormExecutor out of Supervisor, rename executor and supervisor variables
use 2 spaces instead of tab
